chanstate: move commitment diff types
Move CommitDiff and its forwarding reference types into chanstate. This
lets the next commitment store facet name pending remote commitment
state without importing channeldb.

Keep forwarding package persistence and commit-diff serialization in
channeldb for now. The aliases preserve existing call sites while the
KV backend code remains in place.
